+++ task/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.4
+
+- Send client id to Taskserver in message header
+  + e.g., "client: info.tangential.task 0.1.4"
+
 # 0.1.3
 
 - Fix bug where BadCertificateException's Trust button was broken